Update store selectors
