--- packages/bierzo-wallet/src/components/Header/selector.ts
@@ -1,24 +1,0 @@
-import { createSelector } from "reselect";
-
-import { ProcessedTx } from "../../logic/transactions/types/BwParser";
-import { getTransactions } from "../../store/notifications/selectors";
-
-export const confirmedTxSelector = createSelector(
-  getTransactions,
-  (txs: readonly ProcessedTx[]) => {
-    const min = Math.min(txs.length, 3);
-
-    return txs.slice(0, min);
-  },
-);
-
-export const lastTxSelector = createSelector(
-  confirmedTxSelector,
-  (txs: readonly ProcessedTx[]) => {
-    if (txs.length === 0) {
-      return undefined;
-    }
-
-    return txs[0];
-  },
-);
